feat: process embed kots release
start to process the embed kots release. the kots release is embed
inside the helmbin binary as a tar.gz file, we need to extract it
and create a config map called kotsadm-application-metadata in the
default namespace, this configmap contains the kots application cr
and is used as means of interface customisation by the kotsadm.