chore: disable builder image generation
this is not mandatory and we are not sure we want to keep developing
this. as the push is failing (as far as i could read due to org level
permissioning) this commit removes the helmvm builder image build and
publish job from the github action.

we can enable it again in the future if we decide this is a goal we want
to pursuit.