feat: introducing plain prompts (#37)
* feat: introducing plain prompts

plain prompts do not leverage survey package. survey package does not
behave well with stdin redirection so it is quite hard to test an
interactive installation. with plain prompts this should be straight
forward to integrate with tools like `expect`.

this commit also adds an end-to-end test that uses plain prompts.

* Update cmd/helmvm/token.go

Co-authored-by: Ethan Mosbaugh <ethan@replicated.com>

* chore: using interfaces for prompt structs

---------

Co-authored-by: Ethan Mosbaugh <ethan@replicated.com>