feat: allow for disabling add-ons
in order to make kots tests possible we need a way of allowing the
disabling of add-ons so they can install their development version
in a running helmvm cluster and validate.

this commit allows users to disable add-ons with the use of the flag
--disable-addon (may be specified multiple times). this commit also
adds a couple of new e2e tests.