feat: using short token instead of k0s token
short tokens are easier and better looking, this pr introduces them
into the regular node join workflow.

we do a request to kotsadm passing in the short token and receive
back the actual token + some other information, we then write down
the token and run the k0s install command.