feat: creates builder endpoint for custom versions
this commit introduces a new builder endpoint to the project. this
builder endpoint receives as a post request a kots release and its
information and builds a custom version of the embedded-cluster binary,
uploading it to an s3 bucket at the end of the process.

this endpoint returns a signed s3 url from where users can download the
custom built version.