feat: publish alternative k0s url in release metadata (#205)
if we are using an alternative k0s binary for a release we need to
expose this as part of the release metadata. without this we can't
make upgrades use the right binary.

the idea here is: if for a given release we need to replace the k0s
binary we will create a branch for it and make sure that the makefile
points to the right location on the branch. we can then create a tag
on such branch and get a release that points to a different location.

for instace: let's suppose we want to override the k0s binary for the
next release, we could add a value for K0S_BINARY_SOURCE_OVERRIDE in
the makefile, push and tag. this is not optimal as we need to remember
to remove the value from K0S_BINARY_SOURCE_OVERRIDE for subsequent
releases hence the idea of making this in a tag.