feat: remove deprecated 'node upgrade' command and metrics
we are not doing upgrades through the command line anymore so this
changes the workflow and this feature must cease to exist. this commit
removes both the 'node upgrade' command (used to replace the running
version of k0s by the one embedded) and the metrics that were generated
by the binary (node upgrades and upgrades).

the metrics should be generated now by the embedded-cluster-operator
as the autopilot plan moves forward.

squash