feat: apply config overrides during node join
starts to apply patches to both k0s' config `api` and `storage`
properties. patches are received from the kots api and applied
in the following order: embedded overrides -> user provided.

this commit also adds unit tests for the new functionality and
to the old pathing functions.

this is a overview of the changes present on this commit:

- added error treatment in case of failure on override apply.
- removed some empty lines.
- added two new fields to the join token response struct.
- created function to apply overrides in place (/etc/k0s/k0s.yaml).
- add unit tests for multiple override apply functions.
- create an abstraction to make reading embed data easier.