feat: add script to getting join token
adds a script to be used in our e2e tests. this script generates a join
token for a controller join token.

```bash
scripts: ./generate-controller-join-token.js 192.168.86.6
opening a new tab
acessing kotsadm on port 30000
waiting and clickin on the 'Continue to Setup' button
waiting and clicking on 'Advanced' to move on with the certificate
waiting and clicking on 'Proceed' to move on with the certificate
going to the /tls endpoint
waiting and clicking on 'Continue'
waiting and clicking in the password field
typing the password
clicking in the Log in button
waiting and clicking in the Cluster Management tab
waiting and clicking in the Add node button
waiting and clicking in the controller role
waiting and fetching the node join command
sudo ./ec node join 192.168.86.6:30000 59hKbnRkZmAbFFbkZ8WoOGZX
scripts:
```