feat: create methods to better deal with embedded data (#294)
as part of the effort for the new embedded cluster build service we
gonna need better tooling to deal with kots release data.

for example: we need to find the "embedded cluster config" that was
embedded inside the release so we can check what embedded cluster
version it points to in order to build using the right binary. the way
this code was laid down we could only read this data directly from the
binary (os.Executable()).

this commit introduces some functions that will later be leveraged in
during the custom binary built process.