feat: check sshd config before starting the install (#313)
now that we only allow installations with root the ssh installation
method became a huge problem. we can only install if ssh access is
allowed for the root user using a ssh key.

while we work on moving ourselves out of the ssh based installation
we need to at least provide the users with valid information on how
to work around this issue.

this pr does exactly that, upon finding an "invalid" ssh config this
is what is printed on the terminal:

```
rmarasch@ec:~$ sudo ./embedded-cluster install
PermitRootLogin config is set to 'no' in /etc/ssh/sshd_config
This will prevent embedded-cluster from installing.
You can temporarily enable root login by changing the
PermitRootLogin config to 'without-password' and restarting
the sshd service. Once the installation is finished you can
restore the original configuration.
FATA[0000] ssh root access is not allowed
rmarasch@ec:~$
```