feat: add proxy infra support on e2e tests (#355)
this does not do much at this stage other than laying down the pieces
needed to create an infra behind a proxy. i have used this infra to test
if the installation was sending any telemetry data out of the
environment.

makes sense to land this as eventually we gonna start working on airgap
and proxy support.