Adds a reboot flag to `reset` (#455)
* Adds a reboot flag to `reset`

TL;DR
-----

Simplifies resetting a node iwth a reboot flag

Details
-------

Encourages better behavior by making it easier to reboot a node
after resetting it. Instead of having to run two commands to
assure a clean reset:

```
$ sudo ./embedded-cluster reset
$ sudo reboot
```

you just have to

```
$ sudo ./embedded-cluster reset --reboot
```

to do the same.

My implementation shells out to the `reboot` command which takes
a little bit longer than making the minimum required syscalls
(`sync(2)` and `reboot(2)`) because of the niceties it brings along
like sending a wall message and signaling processes more gracefully.

My first implementation just made the syscalls and it was a lot
faster but I decided it probably makes more sense to play nice. I
could be convinced it's not needed since the node shouldn't be
running any other workloads and the app and cluster are already
nicely taken care ot.

* Addresses feedback from @ajp-io and @laverya

* Fixes syntax

* Removes unneeded prompt