feat: make a copy of the binary under /var/lib/embedded-cluster
as part of the installation we also want to copy the binary to the
default location so we can later on serve it to other nodes wishing
to join the cluster.

by placing it the standard location we can also guarantee it will be
upgraded by the local-artifact-mirror binary everytime a cluster
upgrade is executed.