bug: stop using plan id to index autopilot plans (#138)
plan's id is used internally to verify if a given plan has already been
executed or not. if we create two different plans with the same id
autopilot does not seem to be running the second as it consider it
already finished.

we now start to use a random id and use an annotation to keep track of
the 1 to 1 mapping between installation and autopilot plan.