feat: apply built in overrides before end user overrides (#610)
we should apply the release built in overrides before applying the
overrides we read from the command line. we were only applying the
latter before this patch so this also solves a bug.

this commit also adds unit tests for the built in overrides and an
a validation after the installation e2e test.