chore: disable appamor setup
iirc this is a left over from when we were having problems installing
embedded cluster on the github runnbers. in the end the problem was
related to the network manager service stomping over the calico
interfaces. if you are seeing this commit then it means that these step
was not necessary and got removed from the e2e action.