feat: introduces a new `--cidr` command line flag (#1324)
* chore: hide --service-cidr and --pod-cidr flags

these are going to be replaces by the new --cidr flag.

* feat: add the --cidr flag and some validations

adds the --cidr flag and makes sure it is valid an at least a /16 subnet.

* feat: validate the cidr range is private

return an error if the cidr range is not private.

* feat: create function to split network cidr in half

created function to split network cidr in half. we plan to use half for
the pod network and the other half for the service network.

* feat: persist cidr flag on k0s config

forward down the provided cidr to the k0s config. this commit also
changes the functions that automatically sets the no proxy env var.

* chore: fix proxy unit tests (new default cidr)

the new default cidr is different than the old one so some expectations
need to be updated.

* feat: make --cidr incompatible with --pod-cidr and --service-cidr

users are supposed to choose one or the other. if --cidr is present then
users can't use --pod-cidr or --service-cidr.

* chore: store cidr in runtimeconfig

store cidr in the runtimeconfig.

* chore: regenerate operator manifests

a new field was added to the runtime config spec object.

* feat: change message based on installation argument

if during the installation --cidr was used then the restore message
should inform users to use --cidr instead of --pod-cidr and --service-cidr

this annotation isn't yet in the kots code but it will be as a follow up
to this pr.

* chore: remove network cidr from runtime config

we won't use it anymore.

* feat: rollback the old code to stop presernving cidr in the cluster

we wont be preserving the cidr in the cluster anymore.

* feat: add function to merge adjacent networks

we will need a function to verify if two networks are adjacent to each
other and if they are, we will merge them into a single network.

* feat: use adjacent networks detection during restore

if pod and svc networks are adjacent and with the same size we instruct
the user to use --cidr instead of --pod-cidr and --service-cidr.

* chore: remove old code that is no longer used

this is a remeniscent commit to remove old code that is no longer used

* chore: restore old test code

* chore: added a test to verify custom cidr in no-proxy

no-proxy should contain the custom cidr provided by the user.

* chore: make one of our tests use --cidr

make sure we have a test that uses and validates --cidr is being
properly applied.