feat: add support-bundle command (#1351)
* feat: add support-bundle command

add support-bundle command. this command generates a support bundle
using the in-cluster support bundle definitions and the local host
support bundle definitions.

* feat: disable interactive support-bundle

we just want to generate a support bundle and save it to disk so we
don't need to have the interactive mode enabled.

* chore: support-bundle isn't always a json

we can't rely on parsing the output as json, it might contain error
messages mixed up.