feat: warn if online installation is attempted on air gap
vendor portal informs us (through the embedded channel release struct)
if the binary was downloaded with or without an  airgap bundle.

if the binary was downloaded with an airgap bundle, we should warn the user
we must warn the users if they attempt to install the binary without
passing in the airgap bundle flag.